Moving landing to index file and update admin index
